Longitudial canonical correlation analysis

Current version is implemented for only the linear trajectory.

lcca.linear(
  x,
  y,
  varthresh = 0.95,
  projectthresh = 1,
  method = "Wilks",
  verbose = FALSE
)

Arguments

x

object for input list(x, time, J, I, visit)

projectthresh

(default=1) threshold for the dimension reduction projection in lfpca.

method

(defualt='Wilks') test statistic to be used. "Wilks","Hotelling", "Pillai", or "Roy".

verbose

(default=FALSE) print all details

varthreshold

(default=0.95) threshold to detemined the number of components of lpcs.

timeadjust

(defult=FALSE)

Value

ccor xcv_x0: Longitudinal Canonical vector for the intercept for x xcv_x1: Longitudinal Canonical vector for the slope for x xcv_y0: Longitudinal Canonical vector for the intercept for y xcv_y1: Longitudinal Canonical vector for the slope for y

Examples

set.seed(12345678)
r=0.8
mu = c(0,0,0,0,0,0)
stddev = rep(c(8,4,2),2)
cormatx = diag(1,6,6)
cormatx[1,5] <- r
cormatx[5,1] <- r
covmatx = stddev %*% t(stddev) * cormatx

## Generate scores
xi = mvrnorm(n = 100, mu = mu, Sigma = covmatx, empirical = FALSE)
I=100

## X
visit.X =rpois(I,1)+3
time.X = unlist(lapply(visit.X, function(x) scale(c(0,cumsum(rpois(x-1,1)+1)))))
J.X = sum(visit.X)
xi.X = xi[,1:3]
V.x=144
phix0 = matrix(0,V.x,3); phix0[1:12, 1]<-.1; phix0[1:12 + 12, 2]<-.1; phix0[1:12 + 12*2, 3]<-.1
phix1 = matrix(0,V.x,3); phix1[1:12 + 12*3, 1]<-.1; phix1[1:12 + 12*4, 2]<-.1; phix1[1:12 + 12*5, 3]<-.1
phixw = matrix(0,V.x,3); phixw[1:12 + 12*6, 1]<-.1; phixw[1:12 + 12*7, 2]<-.1; phixw[1:12 + 12*8, 3]<-.1
zeta.X = t(matrix(rnorm(J.X*3), ncol=J.X)*c(8,4,2))*2
X = phix0 %*% t(xi.X[rep(1:I, visit.X),]) + phix1 %*% t(time.X * xi.X[rep(1:I, visit.X),]) + phixw %*% t(zeta.X) + matrix(rnorm(V.x*J.X, 0, .1), V.x, J.X)

## Y
visit.Y=rpois(I,1)+3
time.Y = unlist(lapply(visit.Y, function(x) scale(c(0,cumsum(rpois(x-1,1)+1)))))
K.Y = sum(visit.Y)

V.y=81
phiy0 = matrix(0,V.y,3); phiy0[1:9, 1]<-.1; phiy0[1:9 + 9, 2]<-.1; phiy0[1:9 + 9*2, 3]<-.1
phiy1 = matrix(0,V.y,3); phiy1[1:9 + 9*3, 1]<-.1; phiy1[1:9 + 9*4, 2]<-.1; phiy1[1:9 + 9*5, 3]<-.1
phiyw = matrix(0,V.y,3); phiyw[1:9 + 9*6, 1]<-.1; phiyw[1:9 + 9*7, 2]<-.1; phiyw[1:9 + 9*8, 3]<-.1
zeta.Y = t(matrix(rnorm(K.Y*3), ncol=K.Y)*c(8,4,2))*2
xi.Y = xi[,4:6]
Y = phiy0 %*% t(xi.Y[rep(1:I, visit.Y),]) + phiy1 %*% t(time.Y * xi.Y[rep(1:I, visit.Y),]) + phiyw %*% t(zeta.Y) + matrix(rnorm(V.y*K.Y ,0, .1), V.y, K.Y)

x = list(X=X, time=time.X, I=I, J=sum(visit.X),visit=visit.X)
y = list(X=Y, time=time.Y, I=I, J=sum(visit.Y),visit=visit.Y)
re = lcca.linear(x=x,y=y)
#> Wilks' Lambda, using F-approximation (Rao's F):
#>               stat      approx df1      df2      p.value
#> 1 to 4:  0.3635147 6.913769391  16 281.7023 2.345901e-13
#> 2 to 4:  0.9176709 0.904261906   9 226.4882 5.222746e-01
#> 3 to 4:  0.9749770 0.599311064   4 188.0000 6.635830e-01
#> 4 to 4:  0.9999414 0.005566361   1  95.0000 9.406834e-01
re
#> $tests
#> $tests$id
#> [1] "Wilks"
#> 
#> $tests$stat
#> [1] 0.3635147 0.9176709 0.9749770 0.9999414
#> 
#> $tests$approx
#> [1] 6.913769391 0.904261906 0.599311064 0.005566361
#> 
#> $tests$df1
#> [1] 16  9  4  1
#> 
#> $tests$df2
#> [1] 281.7023 226.4882 188.0000  95.0000
#> 
#> $tests$p.value
#> [1] 2.345901e-13 5.222746e-01 6.635830e-01 9.406834e-01
#> 
#> $tests$r
#> [1] 0.7770923 0.2424396 0.1580058 0.0076544
#> 
#> 
#> $ccor.dim
#> [1] 1
#> 
#> $ccor
#> [1] 0.7770923
#> 
#> $xcv_x0
#>                 [,1]
#>   [1,] -5.240458e-02
#>   [2,] -5.121346e-02
#>   [3,] -5.281954e-02
#>   [4,] -5.250476e-02
#>   [5,] -5.162357e-02
#>   [6,] -5.146379e-02
#>   [7,] -5.309163e-02
#>   [8,] -5.314932e-02
#>   [9,] -5.160490e-02
#>  [10,] -5.122553e-02
#>  [11,] -5.311918e-02
#>  [12,] -5.154821e-02
#>  [13,] -3.697314e-03
#>  [14,] -3.578033e-03
#>  [15,] -3.526340e-03
#>  [16,] -5.019905e-03
#>  [17,] -4.005135e-03
#>  [18,] -4.105053e-03
#>  [19,] -4.087958e-03
#>  [20,] -3.603200e-03
#>  [21,] -4.762419e-03
#>  [22,] -3.327134e-03
#>  [23,] -4.107509e-03
#>  [24,] -4.726038e-03
#>  [25,]  9.177084e-03
#>  [26,]  9.409618e-03
#>  [27,]  9.631838e-03
#>  [28,]  1.019491e-02
#>  [29,]  1.038734e-02
#>  [30,]  9.599668e-03
#>  [31,]  9.933208e-03
#>  [32,]  8.924947e-03
#>  [33,]  1.045960e-02
#>  [34,]  1.017450e-02
#>  [35,]  9.659505e-03
#>  [36,]  9.736108e-03
#>  [37,] -8.583329e-04
#>  [38,] -3.837461e-04
#>  [39,] -1.572281e-03
#>  [40,] -5.510319e-04
#>  [41,] -1.396470e-03
#>  [42,]  4.245864e-04
#>  [43,] -9.530606e-04
#>  [44,] -1.761365e-04
#>  [45,] -2.497254e-04
#>  [46,]  2.985649e-04
#>  [47,]  1.041726e-03
#>  [48,] -6.748502e-04
#>  [49,] -5.239793e-04
#>  [50,] -5.905928e-06
#>  [51,]  3.401799e-04
#>  [52,] -4.313707e-04
#>  [53,] -1.301769e-04
#>  [54,] -8.990692e-04
#>  [55,] -1.087787e-04
#>  [56,] -4.575808e-04
#>  [57,]  6.172353e-04
#>  [58,] -9.927459e-04
#>  [59,] -1.161591e-03
#>  [60,] -1.447779e-04
#>  [61,] -6.160998e-04
#>  [62,]  1.139509e-03
#>  [63,] -1.235321e-03
#>  [64,] -6.992568e-04
#>  [65,] -1.762374e-04
#>  [66,]  1.733414e-03
#>  [67,] -2.174736e-04
#>  [68,] -6.148850e-04
#>  [69,]  5.907092e-04
#>  [70,] -8.222421e-04
#>  [71,]  7.182849e-05
#>  [72,]  1.711529e-03
#>  [73,]  2.051787e-02
#>  [74,]  1.980696e-02
#>  [75,]  2.015651e-02
#>  [76,]  1.948428e-02
#>  [77,]  2.005626e-02
#>  [78,]  2.057709e-02
#>  [79,]  2.177505e-02
#>  [80,]  2.038227e-02
#>  [81,]  2.163555e-02
#>  [82,]  2.081181e-02
#>  [83,]  2.074497e-02
#>  [84,]  2.153904e-02
#>  [85,] -2.293309e-03
#>  [86,] -2.106100e-03
#>  [87,] -1.459963e-03
#>  [88,] -3.571095e-03
#>  [89,] -2.168368e-03
#>  [90,] -1.729366e-03
#>  [91,] -2.911988e-03
#>  [92,] -1.310539e-03
#>  [93,] -3.162737e-03
#>  [94,] -1.479016e-03
#>  [95,] -2.757131e-03
#>  [96,] -1.385267e-03
#>  [97,] -4.876773e-03
#>  [98,] -2.812057e-03
#>  [99,] -2.932701e-03
#> [100,] -3.891947e-03
#> [101,] -1.875466e-03
#> [102,] -3.712810e-03
#> [103,] -2.989562e-03
#> [104,] -4.248513e-03
#> [105,] -3.250476e-03
#> [106,] -2.644392e-03
#> [107,] -2.654437e-03
#> [108,] -1.791128e-03
#> [109,]  5.511383e-04
#> [110,] -7.935502e-04
#> [111,] -8.379491e-04
#> [112,] -7.219404e-04
#> [113,]  1.114718e-03
#> [114,] -2.637650e-04
#> [115,]  9.618236e-04
#> [116,] -8.783366e-04
#> [117,]  1.527783e-03
#> [118,] -9.448495e-04
#> [119,] -1.005080e-04
#> [120,] -3.787867e-04
#> [121,]  8.692029e-04
#> [122,]  3.856886e-04
#> [123,] -6.433996e-04
#> [124,]  4.663477e-04
#> [125,]  8.197149e-04
#> [126,] -2.456516e-04
#> [127,] -1.212015e-04
#> [128,]  5.907830e-05
#> [129,] -8.047157e-04
#> [130,]  7.023068e-04
#> [131,]  5.430844e-04
#> [132,]  2.417548e-04
#> [133,] -1.610971e-03
#> [134,] -8.653542e-04
#> [135,]  3.220282e-04
#> [136,]  4.664442e-04
#> [137,]  8.285327e-04
#> [138,] -7.887992e-04
#> [139,] -1.751198e-03
#> [140,] -4.224397e-04
#> [141,]  8.877716e-05
#> [142,]  8.005640e-04
#> [143,]  5.859152e-04
#> [144,]  5.139042e-04
#> 
#> $xcv_x1
#>                 [,1]
#>   [1,] -7.507811e-04
#>   [2,] -1.572859e-03
#>   [3,] -7.527726e-04
#>   [4,] -7.619821e-04
#>   [5,]  6.504041e-04
#>   [6,]  1.463867e-03
#>   [7,]  5.446627e-05
#>   [8,]  9.539625e-05
#>   [9,] -1.147493e-03
#>  [10,] -8.025226e-04
#>  [11,] -2.341298e-03
#>  [12,] -5.276036e-04
#>  [13,]  7.505225e-04
#>  [14,]  5.788080e-04
#>  [15,] -1.203492e-03
#>  [16,]  6.432623e-04
#>  [17,]  2.904233e-04
#>  [18,]  6.186078e-04
#>  [19,]  9.474854e-04
#>  [20,] -6.447103e-04
#>  [21,]  2.073490e-04
#>  [22,]  1.656986e-04
#>  [23,] -8.091541e-04
#>  [24,] -1.202719e-04
#>  [25,]  1.101125e-03
#>  [26,] -8.558459e-04
#>  [27,]  8.387028e-04
#>  [28,]  6.484049e-05
#>  [29,]  4.880481e-04
#>  [30,] -4.152245e-04
#>  [31,]  4.479064e-04
#>  [32,]  7.282962e-04
#>  [33,] -2.352861e-04
#>  [34,]  1.885000e-03
#>  [35,] -1.527236e-03
#>  [36,]  1.601158e-03
#>  [37,] -5.535664e-02
#>  [38,] -5.446659e-02
#>  [39,] -5.341987e-02
#>  [40,] -5.390930e-02
#>  [41,] -5.349713e-02
#>  [42,] -5.378800e-02
#>  [43,] -5.460786e-02
#>  [44,] -5.388483e-02
#>  [45,] -5.347298e-02
#>  [46,] -5.416128e-02
#>  [47,] -5.493500e-02
#>  [48,] -5.401589e-02
#>  [49,] -5.627733e-03
#>  [50,] -5.708574e-03
#>  [51,] -3.889033e-03
#>  [52,] -5.276997e-03
#>  [53,] -4.211039e-03
#>  [54,] -6.157294e-03
#>  [55,] -1.770175e-03
#>  [56,] -4.725905e-03
#>  [57,] -5.516828e-03
#>  [58,] -3.581518e-03
#>  [59,] -4.729501e-03
#>  [60,] -5.064211e-03
#>  [61,]  9.278862e-03
#>  [62,]  8.734523e-03
#>  [63,]  8.411876e-03
#>  [64,]  1.052318e-02
#>  [65,]  9.766372e-03
#>  [66,]  1.043241e-02
#>  [67,]  1.148311e-02
#>  [68,]  9.378426e-03
#>  [69,]  1.060483e-02
#>  [70,]  9.676407e-03
#>  [71,]  9.579854e-03
#>  [72,]  1.109840e-02
#>  [73,]  8.911571e-03
#>  [74,]  8.256658e-03
#>  [75,]  9.986911e-03
#>  [76,]  7.217593e-03
#>  [77,]  8.292849e-03
#>  [78,]  8.352344e-03
#>  [79,]  8.940095e-03
#>  [80,]  7.813016e-03
#>  [81,]  8.741793e-03
#>  [82,]  9.452739e-03
#>  [83,]  8.025833e-03
#>  [84,]  1.015984e-02
#>  [85,] -3.841968e-03
#>  [86,] -3.785574e-03
#>  [87,] -3.296713e-03
#>  [88,] -4.735503e-03
#>  [89,] -3.062659e-03
#>  [90,] -4.551920e-03
#>  [91,] -4.438180e-03
#>  [92,] -4.608802e-03
#>  [93,] -2.092710e-03
#>  [94,] -4.795350e-03
#>  [95,] -2.970079e-03
#>  [96,] -2.763033e-03
#>  [97,] -8.263363e-03
#>  [98,] -6.487589e-03
#>  [99,] -8.086401e-03
#> [100,] -7.514088e-03
#> [101,] -6.920517e-03
#> [102,] -8.354038e-03
#> [103,] -6.980921e-03
#> [104,] -7.529404e-03
#> [105,] -6.047182e-03
#> [106,] -7.960900e-03
#> [107,] -6.923776e-03
#> [108,] -7.578885e-03
#> [109,]  9.632540e-04
#> [110,] -1.694938e-03
#> [111,] -2.040445e-03
#> [112,] -4.980857e-04
#> [113,]  3.208232e-04
#> [114,]  7.657372e-04
#> [115,]  1.828987e-04
#> [116,]  5.660768e-04
#> [117,] -5.549719e-04
#> [118,] -3.533345e-04
#> [119,] -7.872822e-04
#> [120,]  2.710043e-04
#> [121,]  1.182603e-04
#> [122,] -1.163531e-04
#> [123,]  2.239494e-04
#> [124,] -3.647398e-04
#> [125,] -1.385291e-03
#> [126,]  2.505571e-04
#> [127,]  1.030736e-04
#> [128,]  4.856185e-04
#> [129,]  3.515498e-04
#> [130,] -3.814922e-04
#> [131,]  3.799971e-05
#> [132,] -3.622948e-04
#> [133,] -1.181772e-03
#> [134,]  1.196584e-04
#> [135,] -1.748409e-03
#> [136,]  6.940279e-04
#> [137,]  2.336499e-04
#> [138,] -5.759745e-04
#> [139,] -2.130494e-04
#> [140,] -6.125097e-04
#> [141,] -1.419526e-03
#> [142,] -5.823073e-04
#> [143,] -2.194133e-03
#> [144,]  1.291580e-03
#> 
#> $xcv_y0
#>                [,1]
#>  [1,]  0.0081296799
#>  [2,] -0.0007121821
#>  [3,] -0.0046971487
#>  [4,] -0.0060979262
#>  [5,]  0.0033765720
#>  [6,] -0.0093518843
#>  [7,] -0.0010939603
#>  [8,] -0.0064676004
#>  [9,] -0.0007633369
#> [10,] -0.1360628505
#> [11,] -0.1439140406
#> [12,] -0.1394703811
#> [13,] -0.1467906922
#> [14,] -0.1334887662
#> [15,] -0.1437964509
#> [16,] -0.1327310593
#> [17,] -0.1322796955
#> [18,] -0.1403573404
#> [19,]  0.0499203528
#> [20,]  0.0447663780
#> [21,]  0.0529169043
#> [22,]  0.0496646908
#> [23,]  0.0530201680
#> [24,]  0.0550746144
#> [25,]  0.0515301566
#> [26,]  0.0523687091
#> [27,]  0.0490540908
#> [28,]  0.0033611182
#> [29,]  0.0010679894
#> [30,]  0.0049976316
#> [31,] -0.0049667013
#> [32,] -0.0013752645
#> [33,] -0.0023629180
#> [34,] -0.0030392105
#> [35,] -0.0002020371
#> [36,] -0.0011130012
#> [37,]  0.0037422586
#> [38,]  0.0005803851
#> [39,] -0.0035352086
#> [40,] -0.0013307573
#> [41,]  0.0056856888
#> [42,] -0.0035797176
#> [43,] -0.0063618169
#> [44,] -0.0022546061
#> [45,] -0.0030673160
#> [46,]  0.0008698282
#> [47,] -0.0003527553
#> [48,] -0.0056976938
#> [49,] -0.0012913905
#> [50,]  0.0051752148
#> [51,] -0.0043363013
#> [52,] -0.0012338536
#> [53,] -0.0054863084
#> [54,] -0.0058141976
#> [55,]  0.0242222743
#> [56,]  0.0178410720
#> [57,]  0.0240292814
#> [58,]  0.0199759251
#> [59,]  0.0224672074
#> [60,]  0.0123219022
#> [61,]  0.0202528591
#> [62,]  0.0222041339
#> [63,]  0.0130602136
#> [64,]  0.0911716002
#> [65,]  0.0857195197
#> [66,]  0.0966913558
#> [67,]  0.0898881321
#> [68,]  0.0928409603
#> [69,]  0.0975050878
#> [70,]  0.0951640689
#> [71,]  0.0905931149
#> [72,]  0.0950636233
#> [73,]  0.0158051577
#> [74,]  0.0152536382
#> [75,]  0.0114460203
#> [76,]  0.0153253537
#> [77,]  0.0169586381
#> [78,]  0.0079384420
#> [79,]  0.0098089403
#> [80,]  0.0051102139
#> [81,]  0.0115395775
#> 
#> $xcv_y1
#>                [,1]
#>  [1,] -0.0056960564
#>  [2,]  0.0027124152
#>  [3,] -0.0025278197
#>  [4,]  0.0008639328
#>  [5,]  0.0008118146
#>  [6,] -0.0078314871
#>  [7,]  0.0014535092
#>  [8,] -0.0099575303
#>  [9,] -0.0063555415
#> [10,]  0.0113472290
#> [11,]  0.0122184361
#> [12,] -0.0020498337
#> [13,] -0.0003060191
#> [14,]  0.0028717174
#> [15,]  0.0065755128
#> [16,] -0.0064821738
#> [17,]  0.0078707601
#> [18,] -0.0014681876
#> [19,]  0.0014070242
#> [20,]  0.0015542097
#> [21,] -0.0027652749
#> [22,]  0.0043055865
#> [23,] -0.0011381039
#> [24,] -0.0019029225
#> [25,]  0.0010911985
#> [26,]  0.0030406700
#> [27,]  0.0031524983
#> [28,] -0.0029596656
#> [29,] -0.0075011522
#> [30,] -0.0017339975
#> [31,] -0.0012703328
#> [32,] -0.0016711133
#> [33,] -0.0082931902
#> [34,] -0.0016873109
#> [35,] -0.0033544943
#> [36,] -0.0047838727
#> [37,] -0.1239446099
#> [38,] -0.1329827450
#> [39,] -0.1276773264
#> [40,] -0.1273756574
#> [41,] -0.1294425675
#> [42,] -0.1273186424
#> [43,] -0.1348704603
#> [44,] -0.1304592317
#> [45,] -0.1321602473
#> [46,]  0.0525578713
#> [47,]  0.0539062406
#> [48,]  0.0535783299
#> [49,]  0.0555460352
#> [50,]  0.0586410181
#> [51,]  0.0514615985
#> [52,]  0.0567592613
#> [53,]  0.0578764355
#> [54,]  0.0577939657
#> [55,] -0.0279785412
#> [56,] -0.0276155613
#> [57,] -0.0212924005
#> [58,] -0.0306150263
#> [59,] -0.0309203871
#> [60,] -0.0287975250
#> [61,] -0.0291220940
#> [62,] -0.0318872678
#> [63,] -0.0297629824
#> [64,]  0.0113375016
#> [65,]  0.0042988262
#> [66,]  0.0047984061
#> [67,]  0.0125560896
#> [68,]  0.0008934148
#> [69,]  0.0052566070
#> [70,]  0.0075788399
#> [71,]  0.0084162275
#> [72,]  0.0087641349
#> [73,]  0.0223658083
#> [74,]  0.0268251171
#> [75,]  0.0158752671
#> [76,]  0.0266425126
#> [77,]  0.0191938803
#> [78,]  0.0209776762
#> [79,]  0.0171377535
#> [80,]  0.0147452503
#> [81,]  0.0230731125
#> 
#> $scores
#> $scores$x
#>   [1] -1.033560402 -1.825417558  2.411470232  0.057065829 -1.202058624
#>   [6]  0.807877674  0.837386825 -1.330343997 -1.240722621 -2.671073328
#>  [11]  0.584071316  0.319172316  0.548157656 -0.015175030  0.429559387
#>  [16] -1.623070500  0.207317049 -2.023853045 -0.336319217 -0.484823825
#>  [21]  0.293646023  2.143478921  0.283484831  0.532853970  2.247572489
#>  [26] -0.396915523 -0.449590855 -1.535281369  0.455496575  0.434447381
#>  [31]  0.550949052  1.264365108 -0.117905766 -1.322197623 -0.001990266
#>  [36] -0.169378049  0.221089091  1.523622153 -0.055940800  0.609532557
#>  [41] -0.405083263 -0.563034056  0.662082085  1.893857110 -0.747247774
#>  [46]  0.736157947 -0.513880025 -2.087260185  0.351354950 -2.165879487
#>  [51]  0.382208134 -1.100894694 -0.571719158 -0.709795910  0.192818860
#>  [56] -0.375504359  0.136821268  0.988351178 -0.184143962  0.455601256
#>  [61]  0.829727033  0.236987797 -0.929986042 -0.823152512 -0.142428546
#>  [66]  0.577136117 -0.359060561 -0.448926239  0.046221214  0.912052061
#>  [71]  1.571984499  0.982437444 -0.316074967 -1.126567907 -0.951285076
#>  [76]  1.448258455 -0.096414598  0.902015275 -0.606921232  0.942581940
#>  [81] -1.466091864  1.613133425  0.559548281  1.559845125  0.386193054
#>  [86]  0.040842474 -0.387582380 -0.780522122  1.075027278  1.269216552
#>  [91]  0.036206618  0.229919283 -0.798159392 -0.338457941 -0.832971778
#>  [96]  0.957815797  0.505483301 -0.424563310  0.058141936 -1.213416444
#> 
#> $scores$y
#>   [1] -1.75073124 -1.35519030  2.38190712  0.44784929 -0.81679952  0.64684367
#>   [7]  0.28846733 -0.50962650 -1.71214929 -2.31782618  0.28954627  0.01487763
#>  [13] -1.18306872 -0.88424726  0.59663336 -1.78297933  0.87402454 -1.02907353
#>  [19] -0.77445802 -0.11796074  0.63877484  1.64826569  0.90412172  1.04683246
#>  [25]  1.48248602 -0.25155667 -0.71670287 -1.07010172  0.15519631 -0.44000766
#>  [31]  1.02861275  1.05020140 -0.50197578 -0.18783979  0.08167653 -0.10658151
#>  [37]  1.13451359  0.57369410 -0.04651363  0.09024089 -0.20967284 -0.73192012
#>  [43]  0.82177152  0.83488448 -0.21030837  1.52800560 -0.54197169 -0.74041418
#>  [49]  1.15828192 -2.10699938  0.27387100 -1.36191168 -1.19323922 -0.57866142
#>  [55]  0.57293781 -0.14543200  1.24055918  1.41068527  0.14510590 -0.17731410
#>  [61]  1.19523757  0.01097175 -1.64332903 -0.30787260 -0.27594908 -1.18599330
#>  [67] -1.24313865  0.48658436  0.46120924  0.74867386  1.17962691  1.78443459
#>  [73]  0.08051452 -0.74372113  0.13035530  0.09786416 -0.55570469  1.29417261
#>  [79] -0.33985051  1.04274959 -0.82023247  1.40961328 -0.14206371  1.85369215
#>  [85] -0.36560488 -1.03876157 -0.23390584  0.17417254  1.77720926  1.38028300
#>  [91]  0.70958310 -0.53554467 -1.05338167  0.17704992 -0.12314874 -1.13158691
#>  [97]  1.02689410 -0.49406672 -0.45081650 -2.14385211
#> 
#>